tout est validé sauf la fonctionnalité like à faire

+++ backend/controllers/sauce.js
@@ -2,6 +2,7 @@ const Sauce = require('../models/sauce');
 
 const fs = require('fs');
 const sauce = require('../models/sauce');
+const exp = require('constants');
 
 exports.getAllSauce = (req, res, next) => {
     Sauce.find()
@@ -18,10 +19,10 @@ exports.createSauce = (req, res, next) => {
         userId: req.auth.userId,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-  
+
     sauce.save()
-    .then(() => { res.status(201).json({message: 'Objet enregistré !'})})
-    .catch(error => { res.status(400).json( { error })})
+        .then(() => { res.status(201).json({ message: 'Objet enregistré !' }) })
+        .catch(error => { res.status(400).json({ error }) })
 };
 
 exports.getOneSauce = (req, res, next) => {
@@ -35,38 +36,55 @@ exports.modifySauce = (req, res, next) => {
         ...JSON.parse(req.body.sauce),
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
-  
-    delete sauceObject._userId;
-    Sauce.findOne({_id: req.params.id})
-        .then((sauce) => {
-            if (sauce.userId != req.auth.userId) {
-                res.status(401).json({ message : 'Non autorisé'});
-            } else {
-                Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
-                .then(() => res.status(200).json({message : 'Objet modifié !'}))
+
+    if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Non autorisé' });
+
+    } else {
+        if (req.file) {
+            const filename = sauce.imageUrl.split('/images/')[1];
+
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+                    .catch(error => res.status(401).json({ error }));
+            })
+        } else {
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Objet modifié !' }))
                 .catch(error => res.status(401).json({ error }));
-            }
-        })
-        .catch((error) => {
-            res.status(400).json({ error });
-        });
+        }
+    }
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.findOne({ _id: req.params.id})
-       .then(sauce => {
-           if (sauce.userId != req.auth.userId) {
-               res.status(401).json({message: 'Non autorisé'});
-           } else {
-               const filename = sauce.imageUrl.split('/images/')[1];
-               fs.unlink(`images/${filename}`, () => {
-                   Sauce.deleteOne({_id: req.params.id})
-                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
-                       .catch(error => res.status(401).json({ error }));
-               });
-           }
-       })
-       .catch( error => {
-           res.status(500).json({ error });
-       });
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non autorisé' });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
+
+exports.idontknow = (req, res, next) => {
+
+    console.log(req.body)
+
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+}
